pull

+++ src/main/java/com/sparta/spartaboard/service/PostService.java
@@ -3,6 +3,7 @@ package com.sparta.spartaboard.service;
 import com.sparta.spartaboard.domain.dto.PostRequestDto;
 import com.sparta.spartaboard.domain.dto.PostResponseDto;
 import com.sparta.spartaboard.domain.entity.Post;
+import com.sparta.spartaboard.domain.entity.User;
 import com.sparta.spartaboard.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -10,6 +11,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -21,12 +23,12 @@ public class PostService {
         return ResponseEntity.ok(postRepository.findAllByOrderByCreatedAtDesc().stream().map(PostResponseDto::new).toList());
     }
 
-    public ResponseEntity<PostResponseDto> createPost(PostRequestDto postRequestDto) {
-        Post post = postRequestDto.toEntity();
+    public ResponseEntity<PostResponseDto> createPost(PostRequestDto postRequestDto, User user) {
+        Post post = postRequestDto.toEntity(user);
 
         Post savePost = postRepository.save(post);
 
-        return ResponseEntity.ok(new PostResponseDto(savePost));
+        return ResponseEntity.status(201).body(new PostResponseDto(savePost));
     }
 
     public ResponseEntity<PostResponseDto> getPost(Long id) {
@@ -36,16 +38,24 @@ public class PostService {
     }
 
     @Transactional
-    public ResponseEntity<PostResponseDto> editPost(Long id, PostRequestDto postRequestDto) {
+    public ResponseEntity<PostResponseDto> editPost(Long id, PostRequestDto postRequestDto, User user) {
         Post post = findPost(id);
+
+        if (!Objects.equals(post.getUser().getUsername(), user.getUsername())) {
+            throw new IllegalArgumentException("작성자가 일치하지 않습니다.");
+        }
 
         post.update(postRequestDto);
 
         return ResponseEntity.ok(new PostResponseDto(post));
     }
 
-    public ResponseEntity<String> deletePost(Long id, PostRequestDto postRequestDto) {
+    public ResponseEntity<String> deletePost(Long id, User user) {
         Post post = findPost(id);
+
+        if (!Objects.equals(post.getUser().getUsername(), user.getUsername())) {
+            throw new IllegalArgumentException("작성자가 일치하지 않습니다.");
+        }
 
         postRepository.delete(post);
